+++ src/components/swap/CreateSwap.tsx
@@ -12,11 +12,13 @@ import {
   getQueryParams,
 } from "../../utils/utils";
 import { QUERY_PARAMS } from "../../constants/constants";
+import { ecosystems } from "../navbar/connectWallet/constants";
 import { InputAddressAndFeeRateDetails } from "./InputAddressAndFeeRateDetails";
 import { useBitcoinWallet } from "@gardenfi/wallet-connectors";
 import { useEVMWallet } from "../../hooks/useEVMWallet";
 import { useStarknetWallet } from "../../hooks/useStarknetWallet";
 import { rpcStore } from "../../store/rpcStore";
+import { useSolanaWallet } from "../../hooks/useSolanaWallet";
 
 export const CreateSwap = () => {
   const [loadingDisabled, setLoadingDisabled] = useState(false);
@@ -26,12 +28,14 @@ export const CreateSwap = () => {
   const { account: btcAddress, provider } = useBitcoinWallet();
   const { address } = useEVMWallet();
   const { starknetAddress } = useStarknetWallet();
+  const { solanaAnchorProvider } = useSolanaWallet();
   const {
     assets,
     fetchAndSetBitcoinBalance,
     fetchAndSetEvmBalances,
     fetchAndSetFiatValues,
     fetchAndSetStarknetBalance,
+    fetchAndSetSolanaBalance,
     clearBalances,
   } = assetInfoStore();
 
@@ -61,6 +65,9 @@ export const CreateSwap = () => {
   const { setOpenModal } = modalStore();
 
   const buttonLabel = useMemo(() => {
+    if (needsWalletConnection)
+      return `Connect ${capitalizeChain(needsWalletConnection)} Wallet`;
+
     return error.liquidityError
       ? "Insufficient liquidity"
       : error.insufficientBalanceError
@@ -114,20 +121,17 @@ export const CreateSwap = () => {
   }, [inputAsset, outputAsset]);
 
   const handleConnectWallet = () => {
-    if (needsWalletConnection === "starknet") {
-      setOpenModal(modalNames.connectWallet, {
-        Starknet: true,
-        Bitcoin: false,
-        EVM: false,
-      });
-    }
-    if (needsWalletConnection === "evm") {
-      setOpenModal(modalNames.connectWallet, {
-        EVM: true,
-        Starknet: false,
-        Bitcoin: false,
-      });
-    }
+    if (!needsWalletConnection) return;
+
+    const modalState = Object.values(ecosystems).reduce(
+      (acc, { name }) => {
+        acc[name] = name.toLowerCase() === needsWalletConnection;
+        return acc;
+      },
+      {} as Record<string, boolean>
+    );
+
+    setOpenModal(modalNames.connectWallet, modalState);
   };
 
   useEffect(() => {
@@ -143,6 +147,9 @@ export const CreateSwap = () => {
       }
       if (starknetAddress) {
         await fetchAndSetStarknetBalance(starknetAddress);
+      }
+      if (solanaAnchorProvider) {
+        await fetchAndSetSolanaBalance(solanaAnchorProvider.publicKey);
       }
     };
 
@@ -160,8 +167,10 @@ export const CreateSwap = () => {
     fetchAndSetEvmBalances,
     fetchAndSetBitcoinBalance,
     starknetAddress,
+    solanaAnchorProvider,
     fetchAndSetFiatValues,
     fetchAndSetStarknetBalance,
+    fetchAndSetSolanaBalance,
     clearBalances,
     workingRPCs,
   ]);