+++ src/hooks/useSwap.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useMemo, useRef } from "react";
 import { BTC, swapStore } from "../store/swapStore";
 import { IOType, network } from "../constants/constants";
-import { Asset, Chain, isBitcoin } from "@gardenfi/orderbook";
+import { Asset, Chain, isBitcoin, isSolana } from "@gardenfi/orderbook";
 import debounce from "lodash.debounce";
 import { assetInfoStore } from "../store/assetInfoStore";
 import {
@@ -21,6 +21,7 @@ import { ConnectingWalletStore } from "../store/connectWalletStore";
 import orderInProgressStore from "../store/orderInProgressStore";
 import pendingOrdersStore from "../store/pendingOrdersStore";
 import BigNumber from "bignumber.js";
+import { useSolanaWallet } from "./useSolanaWallet";
 import { formatAmount, getOrderPair } from "../utils/utils";
 // import { useWalletClient } from "wagmi";
 // import { Account } from "viem";
@@ -76,12 +77,13 @@ export const useSwap = () => {
       balances[getOrderPair(inputAsset.chain, inputAsset.tokenAddress)],
     [inputAsset, balances]
   );
+  const { solanaAddress } = useSolanaWallet();
 
   const inputTokenBalance = useMemo(
     () =>
       inputBalance &&
       inputAsset &&
-      (!isStarknet(inputAsset.chain)
+      (!isStarknet(inputAsset.chain) && !isSolana(inputAsset.chain)
         ? formatAmount(
             Number(inputBalance),
             inputAsset.decimals,
@@ -383,6 +385,10 @@ export const useSwap = () => {
         check: (chain: Chain) => isStarknet(chain),
         address: starknetAddress,
       },
+      solana: {
+        check: (chain: Chain) => isSolana(chain),
+        address: solanaAddress,
+      },
     };
 
     for (const [chainKey, { check, address }] of Object.entries(
@@ -394,7 +400,7 @@ export const useSwap = () => {
     }
 
     return null;
-  }, [inputAsset, outputAsset, evmAddress, starknetAddress]);
+  }, [inputAsset, outputAsset, evmAddress, starknetAddress, solanaAddress]);
 
   const handleSwapClick = async () => {
     if (needsWalletConnection) {
@@ -469,7 +475,6 @@ export const useSwap = () => {
         receiveAmount: outputAmountInDecimals,
         additionalData,
       });
-
       if (res.error) {
         if (
           res.error.includes(