+++ src/layout/WalletProviders.tsx
@@ -1,0 +1,34 @@
+import { FC, ReactNode } from "react";
+import { Network } from "@gardenfi/utils";
+import { network } from "../constants/constants";
+import { WagmiProvider } from "wagmi";
+import { BTCWalletProvider } from "@gardenfi/wallet-connectors";
+import { StarknetConfig } from "@starknet-react/core";
+import { SolanaProvider } from "./solana/SolanaProvider.tsx";
+import {
+  starknetChains,
+  starknetProviders,
+  connectors as starknetConnectors,
+} from "./starknet/config";
+import { config } from "./wagmi/config";
+
+interface WalletProviderProps {
+  children: ReactNode;
+}
+
+export const WalletProviders: FC<WalletProviderProps> = ({ children }) => {
+  return (
+    <WagmiProvider config={config}>
+      <BTCWalletProvider network={network as Network} store={localStorage}>
+        <StarknetConfig
+          chains={starknetChains}
+          provider={starknetProviders}
+          connectors={starknetConnectors}
+          autoConnect
+        >
+          <SolanaProvider>{children}</SolanaProvider>
+        </StarknetConfig>
+      </BTCWalletProvider>
+    </WagmiProvider>
+  );
+};