+++ src/pages/index.tsx
@@ -13,19 +13,21 @@ import { GardenProvider } from "@gardenfi/react-hooks";
 import { useWalletClient } from "wagmi";
 import { useAccount } from "@starknet-react/core";
 import { Environment as GardenEnvironment } from "@gardenfi/utils";
+import { useSolanaWallet } from "../hooks/useSolanaWallet";
 import { rpcStore } from "../store/rpcStore";
 import { useEffect } from "react";
 
 function App() {
   const { data: walletClient } = useWalletClient();
   const { account: starknetWallet } = useAccount();
+  const { solanaAnchorProvider } = useSolanaWallet();
   const { fetchAndSetRPCs } = rpcStore();
 
   useEffect(() => {
     fetchAndSetRPCs();
-  // eslint-disable-next-line react-hooks/exhaustive-deps
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-  
+
   return (
     <GardenProvider
       config={{
@@ -39,11 +41,13 @@ function App() {
                 info: import.meta.env.VITE_INFO_URL,
                 evmRelay: import.meta.env.VITE_RELAYER_URL,
                 starknetRelay: import.meta.env.VITE_STARKNET_URL,
+                solanaRelay: import.meta.env.VITE_SOLANA_URL,
               }
             : (network as unknown as GardenEnvironment),
         wallets: {
           evm: walletClient,
           starknet: starknetWallet,
+          solana: solanaAnchorProvider ?? undefined,
         },
       }}
     >